refactor(rpc): documentation and better module organization
